feat(ghe): Set GHEDesigner to use GeoJSON polygons
This change switches the package to use the GeoJSON polygons of the GHE field along with the set_geometry_constraints_bi_rectangle_constrained method ONLY when the system parameter file does not contain ghe_geometric_params.

I did it this way in order to introduce as few breaking changes as possible and to get all of the existing tests to pass. A new test has been added, which runs through the routine with detailed geometry from the GeoJSON.

Resolves https://github.com/NREL/ThermalNetwork/issues/44