feat(geometry): Rotate GeoJSON Polygons when a right angle is found (#52)
* feat(geometry): Rotate GeoJSON Polygons when a right angle is found

This commit adds new methods that rotate the GeoJSON polygon if it finds that the lower left corner of the polygon has a right angle. This way, GHEDesigner will place the grid of boreholes along the site boundaries, which will be more spatially efficient and closer to what people would actually build.

* style(tests): Update unit tests and improve styling

* style(lint): Run the linter over all of the files

---------

Co-authored-by: Matt Mitchell <matt.mitchell@nrel.gov>