Add support for all of GHEDesigner's geometric constraints (#67)
* Add support for all of GHEDesigner's geometric constraints

I saw that we were almost all of the way there after [the recent refactor](https://github.com/NREL/ThermalNetwork/pull/63).

It seemed more straightforward to just implement them all while I was adding back the functionality for using the GeoJSON polygons in the Bi-Rectangle Constrained method that was commented out in the refactor. Now, the GeoJSON polygons will get passed along to any of the autosize methods that require a property boundary (including Rowwise).

I updated the old test that I added to make sure that the autosized_rectangle_constrained_borefield runs successfully but I removed the checking of the outputs as I expect them to be a little different when a different autosizing method is used.

Resolves https://github.com/NREL/ThermalNetwork/issues/48

* Remove unnecessary commented code